Debug print Move Gesture using floats
Now that we use double valuators for the motion vector, let's just debug
print using double values instead of casting to int.

BUG=none
TEST=Add -logverbose 10 to X invocation line in /sbin/xstart.sh
TEST=restart ui
TEST=tail -F /var/log/X.org.log | grep Move
TEST=move finger on touchpad and observe...
     Gesture Move: (0.000000, -1.000000)

Reviewed-on: https://gerrit.chromium.org/gerrit/34618
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>

+++ src/gesture.c
@@ -246,8 +246,8 @@ static void Gesture_Gesture_Ready(void* client_data,
             /* TODO(adlr): handle contact initiated */
             break;
         case kGestureTypeMove:
-            DBG(info, "Gesture Move: (%d, %d)\n",
-                (int)gesture->details.move.dx, (int)gesture->details.move.dy);
+            DBG(info, "Gesture Move: (%f, %f)\n",
+                gesture->details.move.dx, gesture->details.move.dy);
             valuator_mask_set_double(
                 rec->mask, CMT_AXIS_X, gesture->details.move.dx);
             valuator_mask_set_double(
